server/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator>=2.1.0

# Database
supabase==2.3.0
psycopg2-binary==2.9.9

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6

# LLM
httpx>=0.24.0,<0.25.0

# Google Calendar
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# External APIs
googlemaps==4.10.0

# Utilities
python-dotenv==1.0.0
dateparser==1.2.0
pytz==2023.3
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
